==> carl_lehman_la_5/Notebooks/01_Democrat_data_gathering.ipynb <==
# Using Reddit's API for Predicting Comments

In this project, I am attempting to answer the question: Do republicans and democrats discuss different topics? To answer this question, I will scrape posts from both republican and democratic subreddits, and then create several classification models to see if I can predict which subreddit the post came from. If I am able to do so with a high degree of accuracy, then this implies that democrats and republicans discuss different topics, or discuss the same topics but with differing sentiments. I will also look at word histograms of popular words form both subreddits, and perform t-tests between top words that appear in both subreddits to see if there is a significant difference between the mean frequencies of the top words.

### Scraping Thread Info from Reddit.com

import requests
import json
import pandas as pd
import numpy as np
import time
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB, GaussianNB, BernoulliNB
from sklearn.decomposition import TruncatedSVD
from sklearn.grid_search import GridSearchCV
from sklearn.neighbors import KNeighborsClassifier

headers = {'User-agent': 'Carl Lehman Bot 0.1'}

res = requests.get(URL0, headers = headers)

res.status_code

posts = []
after = None
for i in range(40):
    print(i)
    if after == None:
        params = []
    else:
        params = {'after': after}
    URL0 = 'https://www.reddit.com/r/Democrat/.json'
    res = requests.get(URL0, params = params, headers = headers)
    if res.status_code == 200:
        data = res.json()
        posts.extend(data['data']['children'])
        after = data['data']['after']
    else:
        print(res.status_code)
        break
    time.sleep(3)

len(posts)

### Putting my posts into a dataframe

democrat = pd.DataFrame([post['data']['selftext'] for post in posts], index = [post['data']['name'] for post in posts])

democrat['title'] = ([post['data']['title'] for post in posts])

democrat['subreddit'] = 0

### Saving my results as a JSON and CSV

# Export to csv
with open('../Data/data_democrat_8_29.json', 'w+') as f:
    json.dump(posts, f)

democrat.to_csv('../Data/democrat.csv')
