==> carl_lehman_la_5/Notebooks/03_Pre-proccessing.ipynb <==
In this notebook I will pre-process my data by creating one dataframe out of the republican and democratic subreddit posts. I will also tokenize my data by count vectorizing and add columns for sentiment analysis to my tokenized matrix. Lastly, I will train test split my reddit dataframe and my term matrix with sentiment analysis in order to be used for modeling.

import requests
import time
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
import csv
import seaborn as sns
import pickle
import matplotlib.pyplot as plt
from sklearn.feature_extraction import stop_words
from scipy.stats import ttest_ind
from nltk.sentiment.vader import SentimentIntensityAnalyzer

% matplotlib inline

# Reading in my data

republican = pd.read_csv('../Data/republican.csv')
democrat = pd.read_csv('../Data/democrat.csv')

Setting my columns

democrat.columns = ['name','text','title','subreddit']
republican.columns = ['name','text','title','subreddit']

### Undersampling my democrat data in order to balance my two classes

republican.shape

democrat.shape

democrat = democrat.sample(n=710, random_state=42)

democrat.shape

republican.shape

### Creating one data frame from my two subreddits

reddit = pd.concat([republican,democrat])
reddit.fillna('', inplace=True)

reddit.shape

Creating a column that contains both the title and body of the post

reddit['text_title'] = reddit['title'] + reddit['text']

### Saving my data frame

reddit.to_csv('../Data/reddit.csv')

### Train test split and save

y = reddit['subreddit']

X = reddit['text_title']

X_train, X_test, y_train, y_test = train_test_split(X,y,stratify=y)

X_train.to_csv('../Data/X_train.csv')

X_test.to_csv('../Data/X_test.csv')

y_train.to_csv('../Data/y_train.csv')

y_test.to_csv('../Data/y_test.csv')

### Creating a count vectorizer data frame

y = list(reddit['subreddit'])

Creating a custom stop word list that includes some numbers and words found in urls

custom_stopwords = list(stop_words.ENGLISH_STOP_WORDS)
custom_stopwords.extend(['10', '12', '13', '14', '15', '18', '25','200', '000','https', 'com', 'youtube', 'www'])

cvec = CountVectorizer(stop_words = custom_stopwords, min_df = 5, max_df = .9)

X_cvec = cvec.fit_transform(X)

term_mat = pd.DataFrame(X_cvec.toarray(), columns=cvec.get_feature_names())

term_mat.head()

### Instantiating my sentiment analyzer

sid = SentimentIntensityAnalyzer()

### Adding columns for sentiment to my data frame

Each post will be given a percent score for each sentiment column, in the hopes that this will help my models differentiate between the subreddits. This is becuase I assume that republicans and democrats will speak with different tones for a given topic or news headline

reddit['positive'] = reddit['text_title'].map(lambda x: sid.polarity_scores(x)['pos'])

reddit['negative'] = reddit['text_title'].map(lambda x: sid.polarity_scores(x)['neg'])

reddit['neutral'] = reddit['text_title'].map(lambda x: sid.polarity_scores(x)['neu'])

### Adding sentiment columns to my term matrix

reddit.reset_index(inplace=True)

reddit.drop(columns='index', inplace=True)

term_mat_sen=term_mat

term_mat_sen['positive'] = reddit['positive']

term_mat_sen['negative'] = reddit['negative']

term_mat_sen['neutral'] = reddit['neutral']

### Train, test split and save my new sentiment term matrix

y = reddit['subreddit']

X_train_sen, X_test_sen, y_train_sen, y_test_sen = train_test_split(term_mat_sen, y, stratify=y)

X_train_sen.to_csv('../Data/X_train_sen.csv')

X_test_sen.to_csv('../Data/X_test_sen.csv')

y_train_sen.to_csv('../Data/y_train_sen.csv')

y_test_sen.to_csv('../Data/y_test_sen.csv')

term_mat.drop(columns=['positive','negative','neutral'],inplace=True)

term_mat['new_target'] = y
term_mat.to_csv('../Data/term_mat.csv')

In my next notebook, EDA, I will make use of this term matrix to show the most prevalent words for Democrats and Republicans through some bar plots. I will also use my train test splitted data for modeling
