==> carl_lehman_la_5/Notebooks/05_modeling_logistic_regression.ipynb <==
In this notebook, I will start my modeling process with a simple logistic regression model. Logistic regression is a good place to start because it is one of the most widely used classification models and is easy to interpret.

import pandas as pd
import numpy as np
import time
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import Pipeline
from sklearn.grid_search import GridSearchCV
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction import stop_words

### Loading in my train test splitted dataframes

reddit = pd.read_csv('../Data/reddit.csv', index_col=0)
X_train = pd.read_csv('../Data/X_train.csv', header = None, index_col=0)
X_test = pd.read_csv('../Data/X_test.csv', header = None, index_col=0)
y_train = pd.read_csv('../Data/y_train.csv', header = None, index_col=0)
y_test = pd.read_csv('../Data/y_test.csv', header = None, index_col=0)
X_train_sen = pd.read_csv('../Data/X_train_sen.csv', index_col=0)
X_test_sen = pd.read_csv('../Data/X_test_sen.csv', index_col=0)
y_train_sen = pd.read_csv('../Data/y_train_sen.csv', header = None, index_col=0)
y_test_sen = pd.read_csv('../Data/y_test_sen.csv', header = None, index_col=0)
reddit.fillna('', inplace=True)

Creating a a custom stopwords list that includes some numbers and terms found in urls

custom_stopwords = list(stop_words.ENGLISH_STOP_WORDS)
custom_stopwords.extend(['10', '12', '13', '14', '15', '18', '25','200', '000','https', 'com', 'youtube', 'www'])

### Logistic Regression on a TFIDF dataframe

I'm gonna start my modeling by tokenzing my data with a tfidf vectorizer, which gives a ratio for each word as opposed to a count, and then fit my logistic regression model on this tokenized data

pipe = Pipeline([
    ('tfidf', TfidfVectorizer(stop_words=custom_stopwords)),
    ('logreg', LogisticRegression())
])

After extensive grid searching, I found these parameters  to be the best.

tfidf_min_df: ignore terms that don't appear in at least a minimum of 1 documents

tfidf_max_df: ignore terms that appear in over 70% of the documents

tfidf_norm: regularizes with the l2 norm

logreg_c: inverse of a regularization strength of 2

logreg_penalty: regularizes with the l1 norm

params = {
    'tfidf__min_df': [1,3,5,7,10],
    'tfidf__max_df': [.7,.75,.8],
    'tfidf__norm': ['l1','l2'],
    'logreg__C': [1,2,4],
    'logreg__penalty': ['l1', 'l2']
}

gs = GridSearchCV(pipe, params)

gs.fit(X_train[1], y_train[1])

### Scoring my model

gs.score(X_train[1],y_train[1])

gs.score(X_test[1],y_test[1])

My model did not perform too well. It appears to be somewhat overfit but not too badly

gs.best_params_

### Logisitic Regression on a count vectorized data frame

count vectorized data is simply just a count of how many times a word appears in the document

pipe = Pipeline([
    ('cvec', CountVectorizer(stop_words=custom_stopwords)),
    ('logreg', LogisticRegression()),
])

After extensive grid searching, I found these parameters  to be the best.

cvec_min_df: ignore terms that don't appear in at least a minimum of 3 documents

cvec_max_df: ignore terms that appear in over 85% of the documents

cvec_ngram_range: the best n gram range was 1, so it only considers each word by itself as opposed to two words

logreg_c: inverse of a regularization strength of .1

logreg_penalty: regularizes with the l2 norm

params = {
    'cvec__min_df': [1,3,5,7,10],
    'cvec__max_df': [.85,.9],
    'cvec__ngram_range': [(1,1),(1,2)],
    'logreg__C': [1,2,4,8,(1/10),(1/5)],
    'logreg__penalty': ['l1', 'l2']
}

gs = GridSearchCV(pipe, params)

gs.fit(X_train[1], y_train[1])

### Scoring my model

gs.score(X_train[1],y_train[1])

gs.score(X_test[1],y_test[1])

My model did not perform too well and seems more overfit than my initial model

gs.best_params_

### Logistic Regression on count vectorized data with sentiment analysis

Let's see if adding features for sentiment analysis will help my logistic regression model

params = {
    'C': [1,2,4,8,(1/10),(1/5)],
    'penalty': ['l1', 'l2']
}

gs = GridSearchCV(LogisticRegression(), params)

data was tokenized via count vectorizing in pre-processing notebook

gs.fit(X_train_sen,y_train_sen[1])

gs.score(X_train_sen,y_train_sen[1])

gs.score(X_test_sen,y_test_sen[1])

Unfortunately, adding sentiment analysis did not help my model much, but it is the best of the three. Let's examine the top coefficients

weights = pd.DataFrame(data = gs.best_estimator_.coef_, columns = X_train_sen.columns)

weights_abs = weights.T.abs()

weights_abs.sort_values(ascending=False, by = 0).head()

The most important features for differentiating between the subreddits is Clinton, democratic, and Kavanaugh. This makes sense because we found in our EDA that Clinton is a top word in the democratic subreddit but was not found often on the republican subreddit

### Conclusion

All of my models scored about a $63 \%$ accuracy. This is only $13 \%$ better than a coin flip. Since my model is having a hard time differentiating between the two subreddits, this supports the hypothesis that democrats and republicans discuss the same subjects. We will try another modeling technique known as K-nearest neighbors next, to see if a non-parametric model will have a higher accuracy score.
