==> carl_lehman_la_5/Notebooks/06_modeling_knn.ipynb <==
Similar to my Logistic Regression notebook, in this notebook I will fit a k-nearest neighbor model on count vectorized data, TFIDF data, and count vectorized data with sentiment analysis in the attempts to create a model that can differentiate between the subreddits. If I am successful at doing so, this will provide as evidence that democrats and republicans have a preference for topics to discuss.

import pandas as pd
import numpy as np
import time
from sklearn.pipeline import Pipeline
from sklearn.grid_search import GridSearchCV
from sklearn.neighbors import KNeighborsClassifier
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction import stop_words

### Loading in my train, test splitted data

reddit = pd.read_csv('../Data/reddit.csv', index_col=0)
X_train = pd.read_csv('../Data/X_train.csv', header = None, index_col=0)
X_test = pd.read_csv('../Data/X_test.csv', header = None, index_col=0)
y_train = pd.read_csv('../Data/y_train.csv', header = None, index_col=0)
y_test = pd.read_csv('../Data/y_test.csv', header = None, index_col=0)
X_train_sen = pd.read_csv('../Data/X_train_sen.csv', index_col=0)
X_test_sen = pd.read_csv('../Data/X_test_sen.csv', index_col=0)
y_train_sen = pd.read_csv('../Data/y_train_sen.csv', header = None, index_col=0)
y_test_sen = pd.read_csv('../Data/y_test_sen.csv', header = None, index_col=0)
reddit.fillna('', inplace=True)

Creating a list of custom stop words that include numbers and some terms found in urls

custom_stopwords = list(stop_words.ENGLISH_STOP_WORDS)
custom_stopwords.extend(['10', '12', '13', '14', '15', '18', '25','200', '000','https', 'com', 'youtube', 'www'])

### KNN on TFIDF data

TFIDF stands for term frequency inverse document frequency, and it essentially gives a word a ratio for how much it appears in the document. I will tokenize my data using TFIDF, then I will run a k-nearest neighbors classifier model on said data

pipe = Pipeline([
    ('tfidf', TfidfVectorizer(stop_words=custom_stopwords)),
    ('knn', KNeighborsClassifier())
])

After extensive grid searching, these are the best parameters I've found to maximize accuracy:

tfidf_min_df: ignore words that aren't found in at least 1 post

tfidf_max_df: ignore words that are found in over $65 \%$ of the documents

tfidf_norm: Regularize the data with an l1 norm

knn_n_neighbors: 50 of the closet neighbors to a point get to vote on what that point's class is

knn_n_weights: weight each point based on distance

knn_n_metric: measure distance with the manhattan metric

params = {
    'tfidf__min_df': [1,3,5,7],
    'tfidf__max_df': [.65,.70,.75,.85],
    'tfidf__norm': ['l1','l2'],
    'knn__n_neighbors': [40,50,60,70],
    'knn__weights': ['uniform', 'distance'],
    'knn__metric': ['minkowski', 'euclidean', 'manhattan']
}

gs = GridSearchCV(pipe, params)

gs.fit(X_train[1], y_train[1])

### Scoring my model

gs.score(X_train[1],y_train[1])

gs.score(X_test[1],y_test[1])

gs.best_params_

My model did not score better than any of my logistic regression models and is more overfit

### KNN on count vectorized data

Now I'll tokenize my data by using counting the appearance of each tearm in each post. Then I'll fit a k-nearest neighbors model to said data.

pipe = Pipeline([
    ('cvec', CountVectorizer(stop_words=custom_stopwords)),
    ('knn', KNeighborsClassifier())
])

After extensive grid searching I found the best parameters to optimize accuracy would be:

cvec_min_df: Ignore words that don't appear in at least 3 documents

cvec_max_df: Ignore words that appear in over $85 \%$ of the documents

cvec_ngram_range: (1,2) Consider each word individually and pairs of words

knn_n_neighbors: 40 of the closest neighbors to a point get to vote on that point's class

knn_weights: each point has a uniform weight

knn_metric: distance is measured using the minkowski metric

params = {
    'cvec__min_df': [1,3,5],
    'cvec__max_df': [.85,.9,.95,1.0],
    'cvec__ngram_range': [(1,1),(1,2)],
    'knn__n_neighbors': [30,40,50],
    'knn__weights': ['uniform', 'distance'],
    'knn__metric': ['minkowski', 'euclidean', 'manhattan']
}

gs = GridSearchCV(pipe, params)

gs.fit(X_train[1], y_train[1])

### Scoring my model

gs.score(X_train[1],y_train[1])

gs.score(X_test[1],y_test[1])

gs.best_params_

My model scored worse than my intial one with an accuracy of $56.3 \%$. However, it is not overfit

### KNN on count vectorized data with sentiment analysis

Let's see if adding features for sentiment analysis will improve my models accuracy in classifying the subreddits

pipe = Pipeline([
    ('cvec', CountVectorizer(stop_words=custom_stopwords)),
    ('knn', KNeighborsClassifier())
])

After extensive grid searching, I found that the best parameters for my model to be as follows:

n_neighbors: 30 neighbors get to vote on each points class

weights: each point is weighted by distance

metric: distance is measured using the minkowski metric

params = {
    'n_neighbors': [30,40,50],
    'weights': ['uniform', 'distance'],
    'metric': ['minkowski', 'euclidean', 'manhattan']
}

gs = GridSearchCV(KNeighborsClassifier(), params)

gs.fit(X_train_sen,y_train_sen[1])

gs.score(X_train_sen,y_train_sen[1])

gs.score(X_test_sen,y_test_sen[1])

gs.best_params_

My model scored a record low of $54.9 \%$, which is only about $5 \%$ better than a coin flip. This is surprising given that we included sentiment analysis. My model is also severely overfit in this case

### Conclusion

We could not create a k-nearest neighbor model that scored better than $60 \%$ on our accuracy metric. Since our model had a difficiult time differentiating between the two subreddit's posts, this is further evidence that democrats and republicans discuss similar topics. In my next modeling notebook, I will try to fit a random forest model to the same train, test splitted data. Random forest models are an ensemble method, meaning that several models are fit and used to determine classes. Since several models are greater than one model, random forest will assumedly perform better on our seemingly indiscernible data.
