==> carl_lehman_la_5/Notebooks/07_modeling_random_forest.ipynb <==
In this notebook I will fit a random forest model on TFIDF tokenized data, count vectorized data, and count vectorized data with sentiment analysis. This is my last attempt at creating a model that can accurately (accuracy score greater than 75 percent) predict which subreddit a post came from 

import pandas as pd
import numpy as np
import time
from sklearn.pipeline import Pipeline
from sklearn.grid_search import GridSearchCV
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction import stop_words

### Loading in my train, test splitted dataframes

reddit = pd.read_csv('../Data/reddit.csv', index_col=0)
X_train = pd.read_csv('../Data/X_train.csv', header = None, index_col=0)
X_test = pd.read_csv('../Data/X_test.csv', header = None, index_col=0)
y_train = pd.read_csv('../Data/y_train.csv', header = None, index_col=0)
y_test = pd.read_csv('../Data/y_test.csv', header = None, index_col=0)
X_train_sen = pd.read_csv('../Data/X_train_sen.csv', index_col=0)
X_test_sen = pd.read_csv('../Data/X_test_sen.csv', index_col=0)
y_train_sen = pd.read_csv('../Data/y_train_sen.csv', header = None, index_col=0)
y_test_sen = pd.read_csv('../Data/y_test_sen.csv', header = None, index_col=0)
reddit.fillna('', inplace=True)

Creating a custom list of stop words that includes numbers and common terms found in urls

custom_stopwords = list(stop_words.ENGLISH_STOP_WORDS)
custom_stopwords.extend(['10', '12', '13', '14', '15', '18', '25','200', '000','https', 'com', 'youtube', 'www'])

### Random Forest on TFIDF data

After extensive grid searching, the best parameters I found are as follows:

tfidf_min_df: Ignore terms that don't appear in at least 1 document

tfidf_max_df: Ignore terms that appear in over $85 \%$ of the documents

tfidf_norm: Regularize the data with an l2 norm

rf_n_estimators: Create 100 decision trees for voting

rf_max_features: The max features for each decision tree is determined by a log2 scale

rf_min_samples_split: At least 3 samples to split an internal node

rf_min_samples_leaf: At least 1 sample to be at the leaf node

pipe = Pipeline([
    ('tfidf', TfidfVectorizer(stop_words=custom_stopwords)),
    ('rf', RandomForestClassifier())
])

params = {
    'tfidf__min_df': [1,3],
    'tfidf__max_df': [.80,.85],
    'tfidf__norm': ['l1','l2'],
    'rf__n_estimators': [90,100],
    'rf__max_features': ['auto', 'log2', 'sqrt'],
    'rf__min_samples_split': [3,4],
    'rf__min_samples_leaf': [1]
}

gs = GridSearchCV(pipe, params)

gs.fit(X_train[1], y_train[1])

### Scoring my model

gs.score(X_train[1],y_train[1])

gs.score(X_test[1],y_test[1])

gs.best_params_

Our first random forest model scored an accuracy of $61.9\%$ and seems to be severely overfit

### Random Forest on count vectorized data

Now I'll see if a random forest model performs better on data that is represented as a count rather than a ratio

pipe = Pipeline([
    ('cvec', CountVectorizer(stop_words=custom_stopwords)),
    ('rf', RandomForestClassifier())
])

After extensive grid searching, the best parameters I found are as follows:

cvec_min_df: Ignore words that don't appear in at least 1 document

cvec_max_df: Ignore words that appear in over $90\%$ of the documents

rf_n_estimators: Make 100 decision trees for voting

rf_max_features: Scale the max features for each decision tree by a square root

rf_min_samples_split: There needs to be at least 3 samples to split an internal node

rf_min_samples_leaf: There needs to be at least 2 samples at a leaf node

params = {
    'cvec__min_df': [1,3],
    'cvec__max_df': [.85,.9],
    'rf__n_estimators': [90,100],
    'rf__max_features': ['auto', 'log2', 'sqrt'],
    'rf__min_samples_split': [3,4],
    'rf__min_samples_leaf': [1,2]
}

gs_cvec = GridSearchCV(pipe, params)

gs_cvec.fit(X_train[1], y_train[1])

### Scoring my model

gs_cvec.score(X_train[1],y_train[1])

gs_cvec.score(X_test[1],y_test[1])

gs_cvec.best_params_

My random forest model performed better on count vectorized data, with an accuracy of $65.9 \%$. It is also less overfit

### Random forest on count vectorized data with sentiment analysis

Let's see if adding features for sentiment analysis will help our random forest model differentiate between the subreddits

After extensive grid searching, the best parameters I found are as follows:

n_estimators: Create 90 decision trees for voting

max_feature: Each decision tree has a max feature size of the log2 of the total features

min_samples_split: There needs to be at least 4 samples to split an internal node

min_samples_leaf: There needs to be at least 1 sample at a leaf node

params = {
    'n_estimators': [80,90],
    'max_features': ['auto', 'log2', 'sqrt'],
    'min_samples_split': [3,4],
    'min_samples_leaf': [1]
}

gs = GridSearchCV(RandomForestClassifier(), params)

gs.fit(X_train_sen,y_train_sen[1])

### Scoring my model

gs.score(X_train_sen,y_train_sen[1])

gs.score(X_test_sen,y_test_sen[1])

gs.best_params_

My model did not improve with an accuracy of $59.1 \%$, thus sentiment analysis did not help with regards to differentiating between the subreddits. Let's look at the most important features from my random forest model

### Most important features

weights = pd.DataFrame(data = [gs.best_estimator_.feature_importances_], columns = X_train_sen.columns)

weights = weights.abs().T

weights.sort_values(ascending=False, by = 0).head()

It seems that in this model the sentiments are the most important features, but this model only scored $9 \%$ better than a coin flip, leaving me to conclude that they are not very good features.

### Conclusion

Random forest performed the best out of all of my models with its best accuracy score being $65.9 \%$ on a count vectorized dataframe. However, this score is only marginal better than my other two models, logistic regression and k-nearest neighbors, which had accuracy scores of $63 \%$ and $60 \%$ respectively. In general, it was hard to differentiate between the two subreddits, r/democrat, and r/republican, leaving me to conclude that the topics discuss on those two subreddits are laregely the same. I suspect this is becuase a large majority of posts are news headlines, and are thus have the same title post on both of the subreddits. This is a good thing, becuase this implies that neither party has a bias or preference as to what news or topics are to be discussed.
